refactor: src 폴더 구조 단순화 및 정리
