+++ src/components/Board/PostsList.tsx
