--- src/routes/DivWrapper/index.tsx
@@ -1,9 +1,0 @@
-import { StrictMode } from "react";
-import { createRoot } from "react-dom/client";
-import { DivWrapper } from "./screens/DivWrapper";
-
-createRoot(document.getElementById("app") as HTMLElement).render(
-  <StrictMode>
-    <DivWrapper />
-  </StrictMode>,
-);
